requirements.txt
aiogram>=3.0.0
fastapi>=0.95.0
uvicorn[standard]>=0.21.1
python-dotenv>=1.0.0